Avances nuxt 3
