+++ pages/archivo/caracteristica/[id].vue
