+++ pages/archivo/descriptor/[id].vue
