+++ pages/archivo/paises/index.vue
@@ -1,47 +1,24 @@
+<script setup>
+const ubicaciones = ref([]);
+
+definePageMeta({ layout: 'con-buscador' });
+
+const { data, error } = await useAsyncGql('Ubicaciones');
+ubicaciones.value = data.value.ubicaciones;
+</script>
+
 <template>
   <div class="pagina">
-    <Cargador v-if="$fetchState.pending" />
+    <!-- <Cargador v-if="$fetchState.pending" /> -->
 
-    <div v-else-if="$fetchState.error">
+    <!-- <div v-else-if="$fetchState.error">
       <h1 class="error">{{ $fetchState.error.message }}</h1>
-    </div>
+    </div> -->
 
-    <div v-else class="contenido">
+    <div class="contenido">
       <Mapa :datos="ubicaciones" />
     </div>
   </div>
 </template>
-
-<script>
-import { gql } from 'nuxt-graphql-request';
-
-export default {
-  layout: 'conBuscador',
-  data() {
-    return {
-      ubicaciones: [],
-    };
-  },
-
-  async fetch() {
-    const query = gql`
-      query {
-        ubicacion {
-          id
-          nombre
-          lat
-          lon
-        }
-      }
-    `;
-
-    const { ubicacion } = await this.$graphql.principal.request(query);
-
-    if (ubicacion.length) {
-      this.ubicaciones = ubicacion;
-    }
-  },
-};
-</script>
 
 <style lang="scss" scoped></style>
