+++ queries/archivo.gql
@@ -1,0 +1,29 @@
+query ArchivoInicio {
+  paginas(filter: { slug: { _eq: "archivo" } }, limit: 1) {
+    titulo
+    slug
+    descripcion
+    contenido
+    banner {
+      id
+      title
+    }
+  }
+}
+
+query ArchivoMenuBuscador {
+  autores(limit: -1) {
+    id
+    nombre
+    apellido
+  }
+
+  # Número de ciudades asociadas al país NO es igual a 0
+  paises(filter: { ciudades_func: { count: { _neq: 0 } } }) {
+    nombre
+  }
+
+  gestos {
+    nombre
+  }
+}
